+++ shinyapps/flood3wms/server.R
@@ -28,7 +28,19 @@ function(input, output, session) {
     
     # basemap
     output$map <- renderLeaflet({
-        leaflet() %>% addTiles() %>% 
+        leaflet() %>% 
+            addWMSTiles(
+                baseUrl = "https://sgx.geodatenzentrum.de/wms_topplus_open?",
+                layers = "web", group = "TopPlusOpen",
+                options = WMSTileOptions(format = "image/png",
+                                         crs = st_crs("epsg:4326"),
+                                         transparent = TRUE),
+                attribution = paste0("BKG (", strftime(Sys.Date(), "%Y"),
+                                     "), <a href=\"https://sgx.geodatenzentrum",
+                                     ".de/web_public/gdz/datenquellen/Datenque",
+                                     "llen_TopPlusOpen.html\">",
+                                     i18n()$t("data source"), "</a>")
+                ) %>% 
             setMaxBounds(lng1 = min(c(df.coor.afe$lon, df.coor.afr$lon)) - 2, 
                          lat1 = min(c(df.coor.afe$lat, df.coor.afr$lat)) - 2, 
                          lng2 = max(c(df.coor.afe$lon, df.coor.afr$lon)) + 2, 
@@ -178,8 +190,8 @@ function(input, output, session) {
                               fill = TRUE, fillColor = "lightblue", 
                               fillOpacity = 0.6, layerId = "afe")
             l %>% addPolygons(lng = df.coor.afr$lon, lat = df.coor.afr$lat,
-                              label = i18n()$t("Rhine"), color = "blue", weight = 2,
-                              fill = TRUE, fillColor = "lightblue", 
+                              label = i18n()$t("Rhine"), color = "blue",
+                              weight = 2, fill = TRUE, fillColor = "lightblue",
                               fillOpacity = 0.6, layerId = "afr")
         } else {
             if (res$river == "Elbe") {
@@ -191,18 +203,20 @@ function(input, output, session) {
             } else if (res$river == "Rhine") {
                 l %>% removeShape(layerId = c("afe"))
                 l %>% addPolygons(lng = df.coor.afr$lon, lat = df.coor.afr$lat,
-                                  label = i18n()$t("Rhine"), color = "blue", weight = 2,
-                                  fill = TRUE, fillColor = "lightblue", 
-                                  fillOpacity = 0.6, layerId = "afr")
+                                  label = i18n()$t("Rhine"), color = "blue",
+                                  weight = 2, fill = TRUE,
+                                  fillColor = "lightblue", fillOpacity = 0.6,
+                                  layerId = "afr")
             } else {
                 l %>% addPolygons(lng = df.coor.afe$lon, lat = df.coor.afe$lat,
                                   label = "Elbe", color = "blue", weight = 2,
                                   fill = TRUE, fillColor = "lightblue", 
                                   fillOpacity = 0.6, layerId = "afe")
                 l %>% addPolygons(lng = df.coor.afr$lon, lat = df.coor.afr$lat,
-                                  label = i18n()$t("Rhine"), color = "blue", weight = 2,
-                                  fill = TRUE, fillColor = "lightblue", 
-                                  fillOpacity = 0.6, layerId = "afr")
+                                  label = i18n()$t("Rhine"), color = "blue",
+                                  weight = 2, fill = TRUE,
+                                  fillColor = "lightblue", fillOpacity = 0.6,
+                                  layerId = "afr")
             }
         }
     }, priority = 5)